+++ imblearn/metrics/__init__.py
@@ -1,0 +1,17 @@
+"""
+The :mod:`imblearn.metrics` module includes score functions, performance
+metrics and pairwise metrics and distance computations.
+"""
+
+from .classification import sensitivity_specificity_support
+from .classification import sensitivity_score
+from .classification import specificity_score
+from .classification import geometric_mean_score
+from .classification import make_index_balanced_accuracy
+from .classification import classification_report_imbalanced
+
+__all__ = [
+    'sensitivity_specificity_support', 'sensitivity_score',
+    'specificity_score', 'geometric_mean_score',
+    'make_index_balanced_accuracy', 'classification_report_imbalanced'
+]